pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[project]
authors = [{name = "Dylan Freedman", email = "freedmand@gmail.com"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "annoy_fixed>=1.16.3",
  "faiss-cpu>=1.7.4",
  "click>=8.1.3",
  "Flask>=2.2.3",
  "joblib>=1.3.0",
  "openpyxl>=3.1.0",
  "openai>=1.0.0",
  "Pillow>=9.5.0",
  "pypdfium2>=4.5.0",
  "python-dotenv>=1.0.0",
  "numpy<2",
  "tiktoken>=0.3.3",
  "torch>=2.0.0",
  "tqdm>=4.65.0",
  "transformers>=4.27.4",
  "setuptools",
  "flask-cors>=6.0.1",
  "pyqt5>=5.15.11",
  "docling>=2.60.0",
]
description = "A semantic search CLI tool"
name = "semantra"
readme = "README.md"
version = "0.1.12"

[project.urls]
"Bug Tracker" = "https://github.com/freedmand/semantra/issues"
"Homepage" = "https://github.com/freedmand/semantra"
"Repository" = "https://github.com/freedmand/semantra"

[project.scripts]
semantra = "semantra.semantra:main"

[tool.setuptools.packages.find]
include = ["semantra*"]
where = ["src"]

[tool.setuptools.package-data]
semantra = ["client_public/**/*"]
